local_panel/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
gunicorn==21.2.0
psutil==5.9.6
psycopg2-binary==2.9.9
paramiko==3.4.0